Move the round robin supplier to an internal class
